requirements.txt
streamlit==1.30.0
python-dotenv
openai
transformers
tensorflow
torch